Move route guards to its own file
